refactor(skills): separate simple creation from the editor - Restore the form as SimpleCreateSkillPage at /simple-create-skill - Keep editor creation and existing skill edits in EditSkillPage - Remove mode routing and verify both product entry points
